+++ exe/mddo_toolbox.rb
@@ -10,7 +10,7 @@ require_relative 'l1_descr/l1_intf_descr_maker'
 require_relative 'nw_subsets/network_sets_diff'
 require_relative 'reach_test/reach_tester'
 require_relative 'reach_test/reach_result_converter'
-require_relative 'convert_namespace/converter'
+require_relative 'convert_namespace/namespace_converter'
 require_relative 'convert_namespace/layer_filter'
 require_relative 'convert_topology/batfish_converter'
 require_relative 'convert_topology/containerlab_converter'
